test: validate ONNX and DeepFaceLive export contracts
Add SAEHD and AMP DFM acceptance coverage for opset 12 contracts, exact names and output order, NHWC float32 boundaries, inline weights, dynamic face batches, eager-vs-ORT parity, TensorFlow/onnxscript independence, structural fp16 exports, and opt-in private checkpoints at their native resolution.

Pin AMP's graph-dynamic morph input across 0.0/0.25/0.50/0.65/1.0, including an independent concat-slice reference proving floor(32*0.65)=20, and replay the one-input SAEHD and type-2 AMP feed contracts used by DeepFaceLive.

Add official-identical merge contract coverage for model predictor callbacks, mask modes, noise gates, blur/sharpen, color-transfer handoffs, output layout, and multi-face composition. Remove the obsolete AMP export-deferral test.

Update checkpoint regression tests to accept official float-kind target casting while retaining cross-kind rejection, and initialize architecture-test weights so activation comparisons cannot inherit NaNs from torch.empty storage.

Verified: SAEHD 7 passed/1 private skip; AMP 6 passed/1 private skip; merge 7 passed; AMP training 14 passed; focused regressions 284 passed/3 skipped; runtime baseline 9 passed; full smoke 668 passed/15 skipped.

+++ tests/smoke/test_fan_extractor.py
@@ -372,7 +372,9 @@ def test_strict_load_1d_state_outside_whitelist(fan2d):
 
 def test_strict_load_dtype_mismatch(fan2d):
     d = dict(FAN2D_OFFICIAL)
-    d["bn1/bias:0"] = np.zeros((1, 1, 1, 64), dtype=np.float64)
+    # Float-kind checkpoint values follow official target-dtype cast
+    # semantics; a cross-kind integer value must still fail.
+    d["bn1/bias:0"] = np.zeros((1, 1, 1, 64), dtype=np.int32)
     _fail_load(fan2d.model, d, "DTYPE_MISMATCH", "bn1/bias:0")
